auth function

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,24 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+           
+    
+        $tasks=$user->tasks()->orderBy('created_at','desc')->paginate(10);
+            
 
-      
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -47,6 +59,8 @@ class TasksController extends Controller
      */
     public function store(Request $request)
     {
+        
+       
         //
          $request->validate([
             'status' => 'required|max:10',   // 追加
@@ -55,12 +69,21 @@ class TasksController extends Controller
 
         
         $task = new Task;
+        
+         $task->user_id=\Auth::user()->id;
+        
         $task->content = $request->content;
         $task->status=$request->status;
+      
         $task->save();
+        
+        
+       
 
         // トップページへリダイレクトさせる
         return redirect('/');
+        
+        
     }
 
     /**
@@ -75,10 +98,17 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
+
+         if (\Auth::id() === $task->user_id) {
+            
         // メッセージ詳細ビューでそれを表示
         return view('tasks.show', [
             'task' => $task,
         ]);
+        
+        }
+        
+        return redirect('/');
     }
     
  
@@ -93,11 +123,15 @@ class TasksController extends Controller
     {
         //
          $task = Task::findOrFail($id);
-
+         
+          
         // メッセージ編集ビューでそれを表示
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        
+       
+       
     }
 
     /**
@@ -133,12 +167,19 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        //
-         $task = Task::findOrFail($id);
        
-        $task->delete();
+        
+         // idの値で投稿を検索して取得
+       
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+       
         return redirect('/');
     }
 }
